fix: standardized imports from lib directory, consoloditated routes for generation
