+++ packages/backend/src/app/api/auth/email-otp/verify/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
-import { handleApiError } from '../../../../../lib/api-response';
-import { verifyEmailOtp } from '../../../../../lib/auth-service';
+import { handleApiError } from '@/lib/api-response';
+import { verifyEmailOtp } from '@/lib/auth-service';
 
 export async function POST(request: Request) {
   try {
