+++ packages/backend/src/app/api/generation/route.ts
