+++ packages/backend/src/app/api/groups/join/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from 'next/server';
-import { handleApiError } from '../../../../lib/api-response';
-import { joinUserGroup } from '../../../../lib/group-membership-service';
-import { getRequestUserId } from '../../../../lib/request-user';
+import { handleApiError } from '@/lib/api-response';
+import { joinUserGroup } from '@/lib/group-membership-service';
+import { getRequestUserId } from '@/lib/request-user';
 
 export async function POST(request: Request) {
   try {
