+++ packages/backend/src/app/api/groups/route.ts
@@ -1,10 +1,10 @@
 import { NextResponse } from 'next/server';
-import { handleApiError } from '../../../lib/api-response';
+import { handleApiError } from '@/lib/api-response';
 import {
   createUserGroup,
   listUserGroups
-} from '../../../lib/group-membership-service';
-import { getRequestUserId } from '../../../lib/request-user';
+} from '@/lib/group-membership-service';
+import { getRequestUserId } from '@/lib/request-user';
 
 export async function GET(request: Request) {
   try {
