+++ packages/backend/src/app/api/ingredients/[ingredientId]/route.ts
@@ -1,10 +1,10 @@
 import { NextResponse } from 'next/server';
-import { handleApiError } from '../../../../lib/api-response';
+import { handleApiError } from '@/lib/api-response';
 import {
   deleteIngredient,
   updateIngredient
-} from '../../../../lib/ingredient-service';
-import { getRequestUserId } from '../../../../lib/request-user';
+} from '@/lib/ingredient-service';
+import { getRequestUserId } from '@/lib/request-user';
 
 // PATCH and DELETE are scoped by both ingredientId and the caller's user id.
 // That ownership check is what enforces "users manage their own pantry items."
