+++ packages/backend/src/app/api/ingredients/route.ts
@@ -1,10 +1,10 @@
 import { NextResponse } from 'next/server';
-import { handleApiError } from '../../../lib/api-response';
+import { handleApiError } from '@/lib/api-response';
 import {
   createIngredient,
   listIngredients
-} from '../../../lib/ingredient-service';
-import { getRequestUserId } from '../../../lib/request-user';
+} from '@/lib/ingredient-service';
+import { getRequestUserId } from '@/lib/request-user';
 
 // These routes currently treat "ingredients" as a user's pantry items.
 // A future canonical ingredient database should probably use a separate route name.
