+++ packages/backend/src/app/api/profiles/me/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from 'next/server';
-import { handleApiError } from '../../../../lib/api-response';
-import { readProfile } from '../../../../lib/profile-service';
-import { getRequestUserId } from '../../../../lib/request-user';
+import { handleApiError } from '@/lib/api-response';
+import { readProfile } from '@/lib/profile-service';
+import { getRequestUserId } from '@/lib/request-user';
 
 export async function GET(request: Request) {
   try {
