--- packages/backend/src/app/profile/constraints/route.ts
@@ -1,1 +1,0 @@
-export { GET, PATCH, POST } from '@/lib/profile-constraints-route';
